% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxcalc-package.R
\docType{package}
\name{toxcalc-package}
\alias{toxcalc-package}
\title{toxcalc: US EPA Whole Effluent Toxicity Statistical Methods}
\description{
An open-source implementation of the statistical methods specified in the US Environmental Protection Agency Whole Effluent Toxicity (WET) method manuals, recreating the capability of the discontinued ToxCalc software. Provides the EPA hypothesis-testing flowchart for deriving no-observed-effect and lowest-observed-effect concentrations, together with the minimum significant difference, and the point-estimation flowchart for median lethal concentrations and inhibition concentrations. Every analysis carries an audit trail recording which test was selected and which manual section justifies the choice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beckyfisher.github.io/toxcalc}
  \item \url{https://github.com/beckyfisher/toxcalc}
  \item Report bugs at \url{https://github.com/beckyfisher/toxcalc/issues}
}

}
\author{
\strong{Maintainer}: Rebecca Fisher \email{r.fisher@aims.gov.au}

Authors:
\itemize{
  \item Rebecca Fisher \email{r.fisher@aims.gov.au}
}

}
\keyword{internal}
