% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxstats-package.R
\docType{package}
\name{toxstats-package}
\alias{toxstats}
\alias{toxstats-package}
\title{toxstats: Standard Endpoint Statistics for Toxicity Test Data}
\description{
Computes the standard endpoints of a concentration-response toxicity test: the no- and lowest-observed-effect concentrations, the minimum significant difference and its percentage of the control, median lethal concentrations, and inhibition concentrations. The method used is selected by the decision rules set out in the US EPA Whole Effluent Toxicity method manuals, and every analysis carries a record of which test was selected and which section of the manual justifies the choice. Those rules are written for a regulatory setting but apply to any concentration-response toxicity data with a control. Recreates the capability of the withdrawn ToxCalc software (Tidepool Scientific), with which this package is not affiliated.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://open-aims.github.io/toxstats}
  \item \url{https://github.com/open-AIMS/toxstats}
  \item Report bugs at \url{https://github.com/open-AIMS/toxstats/issues}
}

}
\author{
\strong{Maintainer}: Rebecca Fisher \email{r.fisher@aims.gov.au}

Authors:
\itemize{
  \item Rebecca Fisher \email{r.fisher@aims.gov.au}
}

}
\keyword{internal}
